feat: card generation
