+++ src/types/FormProps.ts
@@ -1,0 +1,5 @@
+import { FormValues } from './FormValues';
+
+export interface FormProps {
+  setFormValues: React.Dispatch<React.SetStateAction<FormValues[]>>;
+}
