+++ src/types/FormValues.ts
@@ -1,0 +1,7 @@
+export type FormValues = {
+  title?: string;
+  date?: string;
+  role?: string;
+  type?: string;
+  image?: string;
+};
